Implement preview mode: run sketches without installing them.
Uses the same class loader approach as the watch face target. We have made a new module called "sketch-preview", which is the app that loads and displays the sketch. This app isn't built by Android Studio and distributed, however; it is built by APDE on the fly. Part of the build is done on the PC and packaged into the asset "preview_build.zip", then that archive is extracted on the device and the build is finished when needed. There are two main reasons for this approach: crashes and permissions.

We can't have the sketch running class-loaded directly in APDE. If the sketch crashes, then there is no way to prevent the entirety of APDE from crashing. This means that users can't get the stack traces from their sketches' crashes. Further, if APDE crashes, then the report gets sent to Google Play and clutters the actual APDE crashes, which is no good. The solution is to run the sketches in a separate "sketch previewer" app so that if the sketch crashes, that app crashes and not APDE.

Sketches can use arbitrary permissions. If the sketch is run directly in APDE, then APDE needs to declare all of these permissions ahead of time because permissions cannot be dynamically added to the manifest. Some permissions simply cannot be added to APDE because there are privacy concerns - microphone, contacts, camera, etc. This is why we rebuild the sketch previewer. We also can't give all the permissions to the sketch previewer by default for similar privacy reasons. (Also, there could be proprietary/non-built-in permissions that APDE doesn't know about but the sketch wants.) When the user runs a sketch that requires new permissions (e.g. bluetooth), the sketch previewer is rebuilt and reinstalled with the new permission. So the user only needs to install the previewer the first time that they use the new permission, then any sketch using that permission or any others that have been added in the past can be run again without having to reinstall. I think that this is the best compromise between privacy/security concerns and ease of use.

Preview mode is implemented as a new component target, alongside app, wallpaper, watch face, and VR. It is now first in the target list (although not default - perhaps should change that) and the idea is that almost all runs will be in preview mode. Preview build just compiles and dexes the code; it doesn't need to run dx merge or build the APK because the class loader is capable of working with straight dex files and all of the libraries can be added directly to the sketch previewer. Most importantly, the user no longer has to install the sketch each time. On my phone the whole process - from pressing run to seeing the sketch running - takes about two seconds. This is a huge improvement. I don't think I can overstate how significant this is.

I have tested this process extensively on my phone, and it is very fluid, but I have yet to test on other devices or emulators. I am sure that issues will come up.

There are tons of little implementation details, but I think the code (with comments) should speak for itself. I also don't feel like writing it all up right now.

+++ APDE/src/main/java/com/calsignlabs/apde/APDE.java
@@ -1172,15 +1172,17 @@ public class APDE extends Application {
 	 * @throws IOException
 	 */
     public static void deleteFile(File file) throws IOException {
-    	if (file.isDirectory()) {
-    		for (File content : file.listFiles()) {
-    			deleteFile(content);
-    		}
-    	}
-    	
-    	if (!file.delete()) { //Uh-oh...
-    		throw new FileNotFoundException("Failed to delete file: " + file);
-    	}
+    	if (file.exists()) {
+			if (file.isDirectory()) {
+				for (File content : file.listFiles()) {
+					deleteFile(content);
+				}
+			}
+		
+			if (!file.delete()) { //Uh-oh...
+				throw new FileNotFoundException("Failed to delete file: " + file);
+			}
+		}
     }
 	
 	/**
